+++ example/lib/main_advanced.dart
@@ -49,8 +49,7 @@ class _AppViewState extends State<AppView> {
               ),
             ],
             currentIndex: _appRoute.index,
-            onTap: (index) =>
-                setState(() => _appRoute = AppRoute.values[index]),
+            onTap: (index) => setState(() => _appRoute = AppRoute.values[index]),
           ),
         ),
       );